requirements.txt
使用版本：
python==3.10.6
cuda==11.6
pytorch==1.12.1
numpy==1.23.4
torchvision==0.13.1